requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp==3.10.5
opencv-python-headless==4.10.0.84
Pillow>=10.4.0
numpy>=2.0.0
pydub==0.25.1
librosa>=0.10.2
soundfile>=0.12.1
httpx==0.27.2
psutil>=6.0.0
